빈 배열이 리턴되도록 구현
