+++ app/src/test/java/com/codesoom/assignment/application/ProductServiceTest.java
@@ -1,12 +1,16 @@
 package com.codesoom.assignment.application;
 
 import com.codesoom.assignment.domain.entity.ProductRepository;
+import com.codesoom.assignment.models.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import static org.junit.jupiter.api.Assertions.*;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @DisplayName("ProductServiceImpl 에서")
@@ -22,5 +26,17 @@ class ProductServiceImplTest {
     @DisplayName("모든 Product 객체를 불러올 때")
     class Describe_of_readAll_products {
 
+        @Nested
+        @DisplayName("Product 객체가 없을 경우")
+        class Context_without_product {
+
+            @Test
+            @DisplayName("빈 배열을 리턴한다")
+            void it_returns_emtpy_list() {
+                List<Product> products = productService.getProductList();
+
+                assertThat(products).isEmpty();
+            }
+        }
     }
 }
